sync changes from 1.9

+++ Pokecube Core/src/main/java/pokecube/core/entity/pokemobs/helper/EntitySexedPokemob.java
@@ -403,9 +403,9 @@ public abstract class EntitySexedPokemob extends EntityStatsPokemob
             double d1 = this.rand.nextGaussian() * 0.02D;
             double d2 = this.rand.nextGaussian() * 0.02D;
             this.worldObj.spawnParticle(EnumParticleTypes.HEART,
-                    this.posX + (double) (this.rand.nextFloat() * this.width * 2.0F) - (double) this.width,
-                    this.posY + 0.5D + (double) (this.rand.nextFloat() * this.height),
-                    this.posZ + (double) (this.rand.nextFloat() * this.width * 2.0F) - (double) this.width, d0, d1, d2,
+                    this.posX + this.rand.nextFloat() * this.width * 2.0F - this.width,
+                    this.posY + 0.5D + this.rand.nextFloat() * this.height,
+                    this.posZ + this.rand.nextFloat() * this.width * 2.0F - this.width, d0, d1, d2,
                     new int[0]);
         }
